backend/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
celery
redis
pandas
openpyxl
django-cors-headers
drf-spectacular
django-filter
twilio
scikit-learn
python-dotenv
